BREAKING CHANGE: throw errors returned by axe-core's `run` method (#27)
This patch prevents `axe-selenium-java` from swallowing errors returned from `axe-core.run()`. Previously, if axe-core errored during an analysis, the error would not be reported back to the user, but instead, would be swallowed. This would _probably_ cause undefined behavior in people's accessibility tests.

Now, if an error is raised in axe-core, it will be "re-thrown" as an `AXE.AxeRuntimeError`. This allows the caller to handle the error however they see fit. Note that `AXE.AxeRuntimeError` extends `RuntimeException` and does **not require** users to `try/catch` the `analyze()` call which may cause it to be thrown.

Closes #26.